+++ include/commandline/DoubleOption.h
@@ -26,20 +26,28 @@ public:
 		double minimum = -INFINITY, double maximum = INFINITY, int decimals = 1000)
 	: ValidatorOption(names, description, valueName, defaultValue)
 	{
-		setValidator(new QDoubleValidator(minimum, maximum, decimals));
+		_minimum = minimum;
+		_maximum = maximum;
+		setValidator(new QDoubleValidator(_minimum, _maximum, decimals));
 	}
 
 	DoubleOption(const QCommandLineOption &other, double minimum = -INFINITY, double maximum = INFINITY, int decimals = 1000)
 	: ValidatorOption(other)
 	{
-		setValidator(new QDoubleValidator(minimum, maximum, decimals));
+		_minimum = minimum;
+		_maximum = maximum;
+		setValidator(new QDoubleValidator(_minimum, _maximum, decimals));
 	}
 
 	double getDouble(Parser &parser, bool *ok = 0);
 	double *getDoublePtr(Parser &parser, bool *ok = 0);
 
+	bool validate(Parser & parser, QString & value) override;
+
 protected:
 	double _double;
+	int _minimum;
+	int _maximum;
 };
 
 }